backend/pyproject.toml
[project]
name = "recruiting-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic>=2.0.0",
  "python-jose[cryptography]",
  "passlib[bcrypt]",
  "SQLAlchemy>=2.0.0",
  "python-dotenv",
  "openai>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]
